requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
google-generativeai
